requirements.txt
Flask
python-dotenv
trycourier
psycopg2-binary
Flask-JWT-Extended
passlib
bcrypt